requirements.txt
Django==5.1.6
python-decouple==3.8
Requests==2.32.3